+++ Partie 2/Multiplication.java
@@ -11,7 +11,7 @@ public class Multiplication extends Operation {
         super (op1, op2) ;
     }
 
-    public int valeur (){
+    public double valeur (){
         return this.getOperande1().valeur() * this.getOperande2().valeur() ;
     }
